requirements.txt
bcrypt==3.1.7
databases==0.2.6
fastapi==0.52.0
gunicorn==20.0.4
ipdb==0.13.2
ipython==7.13.0
ipython-sql==0.3.9
jupyterlab==2.0.1
passlib==1.7.2
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-multipart==0.0.5
sqlparse==0.3.1
SQLAlchemy==1.3.15
uvicorn==0.11.3
